+++ CHANGELOG.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-This release starts
+Important highlights. This release
 
 - deprecating utilities that duplicate what is available in the Python standard library,
 - renames `birthday.Q` to `birthday.quantile` (old names deprecated)
@@ -28,6 +28,9 @@ This release starts
 ### Added
 
 - `utils.FrozenBidict`
+- `utils.find_zero`
+- `birthday.quantile` (renaming of now deprecated `birthday.Q`)
+- `birthday.probability` (renaming of now deprecated `birthday.P`)
 
 ## 0.5.1 2025-09-01